Move opaque_types::unexpected_hidden_region_diagnostic to error_reporting.
